Move routes folder into a single file
While having a whole routes folder was really great for dependency inversion and
stuff like that, in practice it made it very difficult to know in what order the
routes would be loaded. This gets important as the pages url match is going to be a
wildcard on /. Therefore, we're moving to a routes.js now instead.